implementation of CSRF measures to help further secure phpCollab
    test(Reports, Search, Tasks): Updated tests to match changed code

    feat(CSRF): Added CSRF field to additional views.

    feat(CSRF): Add CSRF field to additional views that may/may not need it.

    feat(Users): Add CSRF field

    feat(Topics): Add CSRF field

    feat(Teams): Add CSRF field

    feat(Tasks): Add CSRF field

    feat(Subtasks): Add CSRF field
        fix(Notifications): Added $logger to instances

    feat(Support): Add CSRF field

    feat(Subtasks): Add CSRF field

    feat(Services): Add CSRF field

    feat(Search): Add CSRF
        Added CSRF field to forms.

    feat(Reports): Add CSRF
        Added CSRF field to forms.

    feat(Projects): Add CSRF
        Added CSRF field to forms.

    feat(Preferences): Add CSRF
        Added CSRF field to forms.

    feat(Phases): Add CSRF
        Added CSRF field to forms.
        fix(Phases): $listPhases was null, so preventing the palette icon for edit phases from working properly
        fix(Phases): Couple bugs wit retrieving data for the editphase view.

    feat(Notes): Add CSRF
        Added CSRF field to forms.

    feat(Newsdesk): Add CSRF
        Added CSRF field to forms.
        fix(Newsdesk): The RSS field was not properly setting when editing a news desk article.  It was setting it to "" instead of 0.
        fix(General): Fixed a bug where the proper "lang".js file was not being loaded for the htmlArea plugin.

    feat(LinkedContent): Add CSRF
        Added CSRF field to forms.
        fix(LinkedContent): There was a bug when adding a peer review, it wasn't passing the projectId or $taskId thru so it was throwing an error.

    feat(Invoices): Add CSRF
        Added CSRF field
        Refactored editinvoiceitem.php to use action method in POST instead of GET.  Also fixed an issue where the selected service was not getting chosen correctly.
        Changed error_log() to $logger object.

    feat(Clients): Add CSRF
        Added CSRF field
        Added additional tests for viewing empty client list, adding a new client, editing a client, and deleting a client.

    feat(Calendar): Add CSRF
        Added CSRF field

    feat(Bookmarks): Add CSRF
        Added CSRF field
        Added additional acceptance tests to test adding bookmark without name or url.
        Moved hard-coded error messages to lang_en.php file

    fix(Logs): Incorrect path for debug logs
        Fixed an issue where /logs/ got changed to /loginlogs/ so errors were not logged in the proper location.

    WIP: passing csrfHandler to openForm method, also changed $logs-> to $loginLogs since it was missed in previous commit.

    WIP: migrating block->openForm to use the csrf handler if it is passed in

    feat(Security): Pull in Symfony CSRF package
        Adding symfony/security-csrf package
        Initializing csrf in library.php

+++ administration/mycompany.php
@@ -40,49 +40,60 @@ if ($session->get('profilSession') != "0") {
 $org = new Organizations();
 
 if ($request->isMethod('post')) {
-    if ($request->request->get("action") == "update") {
+    try {
+        if ($csrfHandler->isValid($request->request->get("csrf_token"))) {
+            if ($request->request->get("action") == "update") {
 
-        $extension = $request->request->get("extension");
-        $extensionOld = $request->request->get("extensionOld");
-        $orgName = $request->request->get("org_name");
-        $orgAddress = $request->request->get("org_address");
-        $orgPhone = $request->request->get("org_phone");
-        $url = $request->request->get("website");
-        $email = $request->request->get("email");
-        $comments = $request->request->get("comments");
-        $logoDel = $request->request->get("logoDel");
+                $extension = $request->request->get("extension");
+                $extensionOld = $request->request->get("extensionOld");
+                $orgName = $request->request->get("org_name");
+                $orgAddress = $request->request->get("org_address");
+                $orgPhone = $request->request->get("org_phone");
+                $url = $request->request->get("website");
+                $email = $request->request->get("email");
+                $comments = $request->request->get("comments");
+                $logoDel = $request->request->get("logoDel");
 
-        if ($logoDel == "on") {
-            $org->setLogoExtensionByOrgId(1, '');
+                if ($logoDel == "on") {
+                    $org->setLogoExtensionByOrgId(1, '');
 
-            try {
-                unlink(APP_ROOT . "/logos_clients/1.$extensionOld");
-            } catch (Exception $e) {
-                echo 'Error deleting file. Message: ' . $e->getMessage();
+                    try {
+                        unlink(APP_ROOT . "/logos_clients/1.$extensionOld");
+                    } catch (Exception $e) {
+                        echo 'Error deleting file. Message: ' . $e->getMessage();
+                    }
+                }
+
+                $extension = strtolower(substr(strrchr($_FILES['logo']['name'], "."), 1));
+
+                try {
+                    if (move_uploaded_file($_FILES['logo']['tmp_name'], "../logos_clients/1.$extension")) {
+                        $org->setLogoExtensionByOrgId(1, $extension);
+                    }
+                } catch (Exception $e) {
+                    echo 'Error moving file. Message: ' . $e->getMessage();
+                }
+
+                $dbParams = [];
+                $dbParams['name'] = phpCollab\Util::convertData($orgName);
+                $dbParams['address1'] = phpCollab\Util::convertData($orgAddress);
+                $dbParams['phone'] = $orgPhone;
+                $dbParams['url'] = $url;
+                $dbParams['email'] = $email;
+                $dbParams['comments'] = phpCollab\Util::convertData($comments);
+
+                $org->updateOrganizationInformation($dbParams);
+
+                phpCollab\Util::headerFunction("../administration/mycompany.php?msg=update");
             }
         }
-
-        $extension = strtolower(substr(strrchr($_FILES['logo']['name'], "."), 1));
-
-        try {
-            if (move_uploaded_file($_FILES['logo']['tmp_name'], "../logos_clients/1.$extension")) {
-                $org->setLogoExtensionByOrgId(1, $extension);
-            }
-        } catch (Exception $e) {
-            echo 'Error moving file. Message: ' . $e->getMessage();
-        }
-
-        $dbParams = [];
-        $dbParams['name'] = phpCollab\Util::convertData($orgName);
-        $dbParams['address1'] = phpCollab\Util::convertData($orgAddress);
-        $dbParams['phone'] = $orgPhone;
-        $dbParams['url'] = $url;
-        $dbParams['email'] = $email;
-        $dbParams['comments'] = phpCollab\Util::convertData($comments);
-
-        $org->updateOrganizationInformation($dbParams);
-
-        phpCollab\Util::headerFunction("../administration/mycompany.php?msg=update");
+    } catch (Exception $e) {
+        $logger->critical('CSRF Token Error', [
+            'edit bookmark' => $request->request->get("id"),
+            '$_SERVER["REMOTE_ADDR"]' => $_SERVER['REMOTE_ADDR'],
+            '$_SERVER["HTTP_X_FORWARDED_FOR"]' => $_SERVER['HTTP_X_FORWARDED_FOR']
+        ]);
+        $msg = 'permissiondenied';
     }
 }
 $company = $org->getOrganizationById(1);
@@ -108,7 +119,8 @@ $block1 = new phpCollab\Block();
 
 echo <<<HTML
 	<form id="{$block1->form}Anchor" method='post' action='../administration/mycompany.php?' name='adminDForm' enctype='multipart/form-data'>
-	<input type='hidden' name='MAX_FILE_SIZE' value='100000000'>
+        <input type="hidden" name="csrf_token" value="{$csrfHandler->getToken()}" />
+	    <input type='hidden' name='MAX_FILE_SIZE' value='100000000'>
 HTML;
